# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Quorum — Multi-tenant Board of Advisors API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      summary: List tenants the current user belongs to
      responses:
        "200":
          description: Tenants list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TenantMembership" }
    post:
      operationId: createTenant
      tags: [tenants]
      summary: Create a new tenant; current user becomes OWNER
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTenantBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tenant" }
  /tenants/{tenantId}/members:
    get:
      operationId: listTenantMembers
      tags: [tenants]
      parameters:
        - $ref: "#/components/parameters/TenantIdParam"
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TenantMember" }
    post:
      operationId: inviteTenantMember
      tags: [tenants]
      parameters:
        - $ref: "#/components/parameters/TenantIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InviteTenantMemberBody" }
      responses:
        "201":
          description: Invited
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TenantMember" }
...
  /boards/{boardId}/members:
    get:
      operationId: listBoardMembers
      tags: [members]
      parameters:
        - $ref: "#/components/parameters/BoardIdParam"
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/BoardMember" }
    post:
      operationId: createBoardMember
      tags: [members]
      parameters:
        - $ref: "#/components/parameters/BoardIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateBoardMemberBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BoardMember" }
  /board-members/{memberId}:
    patch:
      operationId: updateBoardMember
      tags: [members]
      parameters:
        - $ref: "#/components/parameters/MemberIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateBoardMemberBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BoardMember" }
    delete:
      operationId: deleteBoardMember
      tags: [members]
      parameters:
        - $ref: "#/components/parameters/MemberIdParam"
      responses:
        "204":
          description: Deleted
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RequestUploadUrlBody" }
      responses:
        "200":
          description: Presigned URL
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RequestUploadUrlResponse" }
...
components:
  parameters:
    TenantIdParam:
      name: tenantId
      in: path
      required: true
      schema: { type: string }
    BoardIdParam:
      name: boardId
      in: path
      required: true
      schema: { type: string }
    MemberIdParam:
      name: memberId
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    Tenant:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, name, slug, createdAt]
    TenantMembership:
      type: object
      properties:
        tenant: { $ref: "#/components/schemas/Tenant" }
        role: { type: string, enum: [OWNER, ADMIN, EDITOR, VIEWER] }
        joinedAt: { type: string, format: date-time }
      required: [tenant, role, joinedAt]
    TenantMember:
      type: object
      properties:
        userId: { type: string }
        email: { type: ["string", "null"] }
        displayName: { type: ["string", "null"] }
        role: { type: string, enum: [OWNER, ADMIN, EDITOR, VIEWER] }
        joinedAt: { type: string, format: date-time }
      required: [userId, role, joinedAt]
    CreateTenantBody:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        slug: { type: string, minLength: 2, maxLength: 60, pattern: "^[a-z0-9-]+$" }
      required: [name, slug]
    InviteTenantMemberBody:
      type: object
      properties:
        email: { type: string, format: email }
        role: { type: string, enum: [ADMIN, EDITOR, VIEWER] }
      required: [email, role]
...
    BoardMember:
      type: object
      properties:
        id: { type: string }
        boardId: { type: string }
        name: { type: string }
        roleTitle: { type: string }
        lensDescription: { type: ["string", "null"] }
        instructionsText: { type: string }
        groundingDocumentId: { type: ["string", "null"] }
        groundingDocument:
          oneOf:
            - { $ref: "#/components/schemas/GroundingDocument" }
            - { type: "null" }
        modelOverride: { type: ["string", "null"] }
        ordering: { type: integer }
        createdAt: { type: string, format: date-time }
      required: [id, boardId, name, roleTitle, instructionsText, ordering, createdAt]
...
    CreateBoardMemberBody:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        roleTitle: { type: string, minLength: 1, maxLength: 200 }
        lensDescription: { type: ["string", "null"] }
        instructionsText: { type: ["string", "null"] }
        groundingDocumentId: { type: ["string", "null"] }
        modelOverride: { type: ["string", "null"] }
      required: [name, roleTitle]
    UpdateBoardMemberBody:
      type: object
      properties:
        name: { type: ["string", "null"] }
        roleTitle: { type: ["string", "null"] }
        lensDescription: { type: ["string", "null"] }
        instructionsText: { type: ["string", "null"] }
        groundingDocumentId: { type: ["string", "null"] }
        modelOverride: { type: ["string", "null"] }
        ordering: { type: ["integer", "null"] }
    GroundingDocument:
      type: object
      properties:
        id: { type: string }
        tenantId: { type: string }
        filename: { type: string }
        contentType: { type: string }
        storagePath: { type: string }
        characterCount: { type: integer }
        truncated: { type: boolean }
        uploadedAt: { type: string, format: date-time }
      required: [id, tenantId, filename, contentType, storagePath, characterCount, truncated, uploadedAt]
...
    RequestUploadUrlBody:
      type: object
      properties:
        name: { type: string }
        size: { type: integer }
        contentType: { type: string }
      required: [name, size, contentType]
    RequestUploadUrlResponse:
      type: object
      properties:
        uploadURL: { type: string }
        objectPath: { type: string }
      required: [uploadURL, objectPath]
...