# lib/api-zod/src/generated/types/tenantMember.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quorum — Multi-tenant Board of Advisors API
 * OpenAPI spec version: 0.1.0
 */
import type { TenantMemberRole } from "./tenantMemberRole";

export interface TenantMember {
  userId: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  displayName?: string | null;
  role: TenantMemberRole;
  joinedAt: Date;
}